requirements.txt
yfinance
requests
eautifulsoup4
google-generativeai
